refactor: improve module structure for scalability (#57)
