--- src/decorators/capabilities.constants.ts
@@ -1,7 +1,0 @@
-// Prefijo para todas las constantes de metadatos
-export const PREFIX = 'MCP';
-
-// Constantes de metadatos para decoradores de método
-export const MCP_TOOL = `${PREFIX}:tool`;
-export const MCP_PROMPT = `${PREFIX}:prompt`;
-export const MCP_RESOURCE = `${PREFIX}:resource`;
