+++ src/decorators/index.ts
@@ -1,2 +1,5 @@
-export * from './capabilities.constants';
-export * from './capabilities.decorators';
+export * from './prompt.decorator';
+export * from './resolver.decorator';
+export * from './resource.decorator';
+export * from './tool.decorator';
+export * from './user-guard.decorator';
