+++ src/interfaces/index.ts
@@ -1,3 +1,1 @@
-export * from './capabilities.interface';
 export * from './context.interface';
-export * from './mcp-server-options.interface';
