--- src/registry/index.ts
@@ -1,2 +1,0 @@
-export * from './discovery.service';
-export * from './registry.service';
