+++ src/services/index.ts
@@ -1,1 +1,3 @@
+export * from '../services/discovery.service';
+export * from '../services/registry.service';
 export * from './session.manager';
