+++ src/transports/sse/sse.service.spec.ts
@@ -1,20 +1,20 @@
 import { DiscoveryModule } from '@nestjs/core';
 import { Test, TestingModule } from '@nestjs/testing';
 
-import { DiscoveryService } from '../../registry/discovery.service';
-import { McpLoggerService } from '../../registry/logger.service';
-import { RegistryService } from '../../registry/registry.service';
+import { DiscoveryService } from '../../services/discovery.service';
+import { McpLoggerService } from '../../services/logger.service';
+import { RegistryService } from '../../services/registry.service';
 import { SessionManager } from '../../services/session.manager';
-import { StreamableService } from './streamable.service';
+import { SseService } from './sse.service';
 
-describe('StreamableService', () => {
-  let service: StreamableService;
+describe('SseService', () => {
+  let service: SseService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       imports: [DiscoveryModule],
       providers: [
-        StreamableService,
+        SseService,
         McpLoggerService,
         RegistryService,
         SessionManager,
@@ -23,14 +23,10 @@ describe('StreamableService', () => {
           provide: 'MCP_SERVER_OPTIONS',
           useValue: { serverInfo: {}, options: {} },
         },
-        {
-          provide: 'MCP_TRANSPORT_OPTIONS',
-          useValue: { streamable: { options: {} } },
-        },
       ],
     }).compile();
 
-    service = module.get<StreamableService>(StreamableService);
+    service = module.get<SseService>(SseService);
   });
 
   it('should be defined', () => {
